+++ examples/plot_pitch_stability.jl
@@ -27,7 +27,7 @@ set.v_wind = 8.0
 set.cmq = 0.0 # -0.09
 
 # the following values can be changed to match your interest
-dt = 0.05
+dt::Float64 = 0.05
 fg = 2            # cut-off frequency for the filter in Hz
 use_butter  = true
 order = 4         # order of the Butterworth filter
