+++ examples/steering_test_4p.jl
@@ -38,7 +38,7 @@ PRINT = false
 STATISTIC = false
 # end of user parameter section #
 
-dt = 1/set.sample_freq
+dt::Float64 = 1/set.sample_freq
 particles = set.segments + 5
 logger = Logger(particles, STEPS)
